# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 1.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Overseerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, and `/auth/login` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /worker/stats:
    get:
      operationId: getWorkerStats
      tags: [Metrics]
      summary: Get worker stats (alias)
      description: Alias for `GET /metrics/worker`. Returns identical data. Provided for dashboard consumption convenience.
      responses:
        "200":
          description: Worker metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerStats"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ===========================================================================
  # Dashboard & Lifetime Stats
  # ===========================================================================
  /dashboard-stats:
    get:
      operationId: getDashboardStats
      tags: [Dashboard]
      summary: Get aggregated dashboard statistics
      description: |
        Returns combined statistics for the dashboard: lifetime totals (bytes
        reclaimed, items removed, engine runs), current protected count, and
        7-day library growth rate.
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /lifetime-stats:
    get:
      operationId: getLifetimeStats
      tags: [Dashboard]
      summary: Get cumulative lifetime statistics
      description: |
        Returns the cumulative lifetime counters. These persist across container
        restarts and are **not** cleared by the data reset action.
      responses:
        "200":
          description: Lifetime statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifetimeStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Disk Groups
  # ===========================================================================
  /disk-groups:
    get:
      operationId: listDiskGroups
      tags: [Disk Groups]
      summary: List all disk groups
      description: Returns all discovered disk groups (mount points) with their current capacity and threshold settings.
      responses:
        "200":
          description: List of disk groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiskGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /disk-groups/{id}:
    put:
      operationId: updateDiskGroup
      tags: [Disk Groups]
      summary: Update disk group thresholds
      description: |
        Update the threshold and target percentages for a disk group.
        Both values must be between 1 and 99, and threshold must be greater
        than target.
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiskGroupThresholdUpdate"
      responses:
        "200":
          description: Updated disk group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiskGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Disk group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Engine
  # ===========================================================================
  /engine/run:
    post:
      operationId: triggerEngineRun
      tags: [Engine]
      summary: Trigger an engine evaluation
      description: |
        Triggers an immediate evaluation cycle. If a run is already pending
        (queued but not started), returns `"already_pending"` instead.
      responses:
        "200":
          description: Engine run triggered or already pending
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [triggered, already_pending]
                example:
                  status: triggered
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /preferences:
    get:
      operationId: getPreferences
      tags: [Preferences]
      summary: Get preferences
      description: Returns the global preference singleton. If no preferences exist yet, returns defaults.
      responses:
        "200":
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updatePreferences
      tags: [Preferences]
      summary: Update preferences
      description: |
        Updates the global preferences. All weight values must be between 0 and
        10. The log level is applied immediately. If `pollIntervalSeconds` is
        less than 30, it defaults to 300.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferenceSet"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceSet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Rule Fields & Rule Values
  # ===========================================================================
...
  /events:
    get:
      operationId: subscribeEvents
      tags: [Events]
      summary: Subscribe to real-time events via SSE
      description: |
        Opens a long-lived Server-Sent Events (SSE) connection that streams
        typed events in real-time. Events include engine runs, deletions,
        config changes, approval actions, and more (34 event types total).

        Supports reconnection replay via the `Last-Event-ID` header.
        A keepalive comment is sent every 30 seconds to prevent proxy timeouts.
      parameters:
        - name: Last-Event-ID
          in: header
          description: Resume from this event ID on reconnection. Events after this ID will be replayed from a ring buffer (up to 100 events).
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each event follows the SSE format:
                  ```
                  id: <event-id>
                  event: <event-type>
                  data: <json-payload>
                  ```
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ===========================================================================
  # Data
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: integer

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request — invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
...
    DiskGroup:
      type: object
      properties:
        id:
          type: integer
          example: 1
        mountPath:
          type: string
          example: "/media"
        totalBytes:
          type: integer
          format: int64
          example: 1099511627776
        usedBytes:
          type: integer
          format: int64
          example: 879609302221
        thresholdPct:
          type: number
          format: double
          description: Cleanup triggers when usage exceeds this percentage
          example: 85.0
        targetPct:
          type: number
          format: double
          description: Cleanup frees space down to this percentage
          example: 75.0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mountPath, totalBytes, usedBytes, thresholdPct, targetPct]
    DiskGroupThresholdUpdate:
      type: object
      properties:
        thresholdPct:
          type: number
          format: double
          minimum: 1
          maximum: 99
          description: Cleanup triggers when usage exceeds this percentage
          example: 85.0
        targetPct:
          type: number
          format: double
          minimum: 1
          maximum: 99
          description: Cleanup frees space down to this percentage
          example: 75.0
      required: [thresholdPct, targetPct]

    # ---- Integration ----
...
    PreferenceSet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        logLevel:
          type: string
          enum: [debug, info, warn, error]
          default: info
        auditLogRetentionDays:
          type: integer
          description: Days to retain audit logs (0 = forever)
          default: 30
          example: 30
        pollIntervalSeconds:
          type: integer
          description: Polling interval in seconds (minimum 30, default 300)
          minimum: 30
          default: 300
          example: 300
        watchHistoryWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 10
        lastWatchedWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 8
        fileSizeWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 6
        ratingWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 5
        timeInLibraryWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 4
        seriesStatusWeight:
          type: integer
          minimum: 0
          maximum: 10
          default: 3
        executionMode:
          type: string
          enum: [dry-run, approval, auto]
          default: dry-run
        tiebreakerMethod:
          type: string
          enum: [size_desc, size_asc, name_asc, oldest_first, newest_first]
          default: size_desc
        deletionsEnabled:
          type: boolean
          description: Whether actual deletion execution is enabled
          example: false
        updatedAt:
          type: string
          format: date-time
      required:
        - logLevel
        - executionMode
        - tiebreakerMethod

    # ---- Rules ----
...
    WorkerStats:
      type: object
      properties:
        executionMode:
          type: string
          enum: [dry-run, approval, auto]
          example: dry-run
        isRunning:
          type: boolean
          description: Whether the poller is currently executing
        pollIntervalSeconds:
          type: integer
          example: 300
        queueDepth:
          type: integer
          description: Number of items in the deletion queue
        lastRunEvaluated:
          type: integer
          format: int64
          description: Items evaluated in the last (or current) run
        lastRunFlagged:
          type: integer
          format: int64
          description: Items flagged for deletion in the last (or current) run
        lastRunFreedBytes:
          type: integer
          format: int64
          description: Bytes freed in the last (or current) run
        lastRunEpoch:
          type: integer
          format: int64
          description: Unix timestamp of the last run start
        currentlyDeleting:
          type: string
          description: Title of the item currently being deleted (empty if idle)
        protectedCount:
          type: integer
          format: int64
          description: Number of protected items in the last run
        evaluated:
          type: integer
          format: int64
          description: Cumulative total items evaluated across all runs
        actioned:
          type: integer
          format: int64
          description: Cumulative total items actioned (flagged/deleted) across all runs
        freedBytes:
          type: integer
          format: int64
          description: Cumulative total bytes freed across all runs
        processed:
          type: integer
          format: int64
          description: Total items processed by the deletion worker
        failed:
          type: integer
          format: int64
          description: Total items that failed during deletion
      required:
        - executionMode
        - isRunning
        - pollIntervalSeconds
        - queueDepth
        - lastRunEvaluated
        - lastRunFlagged
        - lastRunFreedBytes
        - lastRunEpoch
        - currentlyDeleting
        - protectedCount
        - evaluated
        - actioned
        - freedBytes
        - processed
        - failed

    # ---- Dashboard Stats ----
    DashboardStats:
      type: object
      properties:
        totalBytesReclaimed:
          type: integer
          format: int64
          description: Lifetime total bytes reclaimed
          example: 549755813888
        totalItemsRemoved:
          type: integer
          description: Lifetime total items removed
          example: 42
        totalEngineRuns:
          type: integer
          description: Lifetime total engine evaluation cycles
          example: 150
        protectedCount:
          type: integer
          format: int64
          description: Current number of protected items (from worker metrics)
          example: 12
        growthBytesPerWeek:
          type: integer
          format: int64
          description: Library growth rate in bytes over the last 7 days
          example: 10737418240
        hasGrowthData:
          type: boolean
          description: Whether growth data is available (need at least 7 days of history)
      required:
        - totalBytesReclaimed
        - totalItemsRemoved
        - totalEngineRuns
        - protectedCount
        - growthBytesPerWeek
        - hasGrowthData

    # ---- Lifetime Stats ----
    LifetimeStats:
      type: object
      properties:
        id:
          type: integer
          example: 1
        totalBytesReclaimed:
          type: integer
          format: int64
          example: 549755813888
        totalItemsRemoved:
          type: integer
          example: 42
        totalEngineRuns:
          type: integer
          example: 150
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
      required: [id, totalBytesReclaimed, totalItemsRemoved, totalEngineRuns]

    # ---- Audit ----
...